pyproject.toml
[tool.poetry]
name = "sdk-test"
version = "0.1.0"
description = ""
authors = ["Bond Eng <engineering@bond.tech>"]

[tool.poetry.dependencies]
python = "^3.7"
uvicorn = "^0.12.2"
gunicorn = "^20.0.4"
fastapi = "^0.61.2"
python-dotenv = "^0.15.0"
requests = "^2.25.0"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
